refactor scannetv2
