+++ data/scannetv2.py
@@ -1,0 +1,8 @@
+from .custom import CustomDataset
+
+
+class ScanNetDataset(CustomDataset):
+
+    CLASSES = ('cabinet', 'bed', 'chair', 'sofa', 'table', 'door', 'window', 'bookshelf', 'picture',
+               'counter', 'desk', 'curtain', 'refrigerator', 'shower curtain', 'toilet', 'sink',
+               'bathtub', 'otherfurniture')
